# Dyno: cache identifier lookups to improve resolution performance