… frontend/include/chpl/framework/all-global-strings.h